requirements.txt
tkinter
ttkthemes
pillow
gitpython
